Refactored the massive folder_types.py into separate files and a submodule. All tests are passing.
