+++ python/tank/folder/folder_types/__init__.py
@@ -1,0 +1,23 @@
+# Copyright (c) 2013 Shotgun Software Inc.
+# 
+# CONFIDENTIAL AND PROPRIETARY
+# 
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit 
+# Source Code License included in this distribution package. See LICENSE.
+# By accessing, using, copying or modifying this work you indicate your 
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
+# not expressly granted therein are reserved by Shotgun Software Inc.
+
+"""
+This module contains all the implementations for the different
+folder types that can be created.
+"""
+
+from .errors import EntityLinkTypeMismatch
+from .static import Static
+from .listfield import ListField
+from .entity import Entity
+from .project import Project
+from .user import UserWorkspace
+from .step import ShotgunStep
+from .task import ShotgunTask
